add common error handler
