+++ .vscode/eps-prescription-status-update-api.code-workspace
@@ -37,14 +37,18 @@
       "path": "../packages/checkPrescriptionStatusUpdates"
     },
     {
-      "name": "packages/common",
-      "path": "../packages/common"
+      "name": "packages/common/testing",
+      "path": "../packages/common/testing"
+    },
+    {
+      "name": "packages/common/middyErrorHandler",
+      "path": "../packages/common/middyErrorHandler"
     }
   ],
   "settings": {
     "jest.disabledWorkspaceFolders": [
       "eps-prescription-status-update-monorepo",
-      "packages/common",
+      "packages/common/testing",
       "packages/specification"
     ],
     "files.exclude": {
