+++ packages/checkPrescriptionStatusUpdates/package.json
@@ -20,7 +20,8 @@
     "@aws-sdk/lib-dynamodb": "^3.600.0",
     "@middy/core": "^5.4.0",
     "@middy/http-header-normalizer": "^5.4.1",
-    "@middy/input-output-logger": "^5.4.0"
+    "@middy/input-output-logger": "^5.4.0",
+    "@PrescriptionStatusUpdate_common/middyErrorHandler": "^1.0.0"
   },
   "devDependencies": {
     "@PrescriptionStatusUpdate_common/testing": "^1.0.0"
