+++ packages/common/middyErrorHandler/.eslintrc
@@ -1,0 +1,38 @@
+{
+    "root": true,
+    "env": {
+      "node": true,
+      "es6": true,
+      "jest": true
+    },
+    "ignorePatterns": ["**/lib/*"],
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {"ecmaVersion": "latest"},
+    "plugins": ["import-newlines"],
+    "extends": [
+      "eslint:recommended",
+      "plugin:@typescript-eslint/eslint-recommended",
+      "plugin:@typescript-eslint/recommended"
+    ],
+    "rules": {
+      "block-spacing": "error",
+      "brace-style": ["error", "1tbs"],
+      "comma-dangle": ["error", "never"],
+      "comma-spacing": ["error", {"before": false, "after": true}],
+      "dot-location": ["error", "property"],
+      "eol-last": ["error", "always"],
+      "eqeqeq": "error",
+      "func-call-spacing": "error",
+      "func-style": ["error", "declaration", {"allowArrowFunctions": true}],
+      "import-newlines/enforce": ["error", {"items": 3, "max-len": 120, "semi": false}],
+      "indent": ["error", 2, {"SwitchCase": 1}],
+      "max-len": ["error", 120],
+      "no-multi-spaces": "error",
+      "no-multiple-empty-lines": ["error", {"max": 1}],
+      "no-trailing-spaces": "error",
+      "object-curly-spacing": ["error", "never"],
+      "quotes": ["error", "double", {"allowTemplateLiterals": true, "avoidEscape": true}],
+      "semi": ["error", "never"]
+    }
+  }
+
