+++ packages/common/middyErrorHandler/.vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "jest.jestCommandLine": "/workspaces/eps-prescription-status-update-api/node_modules/.bin/jest --no-cache",
+  "jest.nodeEnv": {
+    "POWERTOOLS_DEV": true,
+    "NODE_OPTIONS": "--experimental-vm-modules"
+  }
+}
