+++ packages/common/middyErrorHandler/jest.config.ts
@@ -1,0 +1,9 @@
+import defaultConfig from "../../../jest.default.config"
+import type {JestConfigWithTsJest} from "ts-jest"
+
+const jestConfig: JestConfigWithTsJest = {
+  ...defaultConfig,
+  rootDir: "./"
+}
+
+export default jestConfig
