+++ packages/common/middyErrorHandler/package.json
@@ -1,0 +1,23 @@
+{
+    "name": "@PrescriptionStatusUpdate_common/middyErrorHandler",
+    "version": "1.0.0",
+    "description": "Common module for a partial middyErrorHandler",
+    "author": "NHS Digital",
+    "license": "MIT",
+    "main": "lib/src/index.js",
+    "type": "module",
+    "scripts": {
+        "unit": "POWERTOOLS_DEV=true NODE_OPTIONS=--experimental-vm-modules jest --no-cache --coverage",
+        "lint": "eslint . --ext .ts --max-warnings 0 --fix",
+        "compile": "tsc",
+        "test": "npm run compile && npm run unit",
+        "check-licenses": "license-checker --failOn GPL --failOn LGPL --start ../.."
+    },
+    "dependencies": {
+        "@aws-lambda-powertools/logger": "^2.2.0"
+    },
+    "devDependencies": {
+        "@middy/core": "^5.4.1",
+        "@PrescriptionStatusUpdate_common/testing": "^1.0.0"
+    }
+}
