+++ packages/common/middyErrorHandler/src/index.ts
@@ -1,0 +1,3 @@
+import {MiddyErrorHandler} from "./errorHandler"
+
+export {MiddyErrorHandler}
