+++ packages/gsul/package.json
@@ -22,7 +22,7 @@
     "@middy/core": "^5.4.0",
     "@middy/input-output-logger": "^5.4.0",
     "@middy/validator": "^5.4.0",
-    "@nhs/fhir-middy-error-handler": "^2.0.8",
+    "@PrescriptionStatusUpdate_common/middyErrorHandler": "^1.0.0",
     "json-schema-to-ts": "^3.1.0"
   }
 }
