+++ tsconfig.build.json
@@ -5,6 +5,7 @@
   // Building this project will build all of the following:
   "references": [
     {"path": "packages/common/testing"},
+    {"path": "packages/common/middyErrorHandler"},
     {"path": "packages/gsul"},
     {"path": "packages/updatePrescriptionStatus"},
     {"path": "packages/sandbox"},
